Refactor post data imports and add posts data file
