+++ fullstack/fullstack/src/app/api/posts/route.js
@@ -1,4 +1,4 @@
-import { posts } from '../../../../data/posts.js';
+import { posts } from '../../data/posts.js';
 import { NextResponse } from 'next/server';
 
 export async function GET() {
