+++ fullstack/fullstack/src/app/data/posts.js
